move list_cmake_dir_optios to util
